requirements.txt
fastapi
uvicorn
pillow
numpy
rembg
python-multipart